+++ CHANGELOG.md
@@ -19,4 +19,5 @@
 - [BeanFactoryUtils returns all beans including beans defined in ancestor bean factories](https://github.com/Tencent/spring-cloud-tencent/pull/517)
 - [fix:add spring-cloud-starter-bootstrap dependency for example](https://github.com/Tencent/spring-cloud-tencent/pull/521)
 - [Feature: Add router actuate endpoint](https://github.com/Tencent/spring-cloud-tencent/pull/523)
+- [Optimize router label resolver spi](https://github.com/Tencent/spring-cloud-tencent/pull/524)
 - [Code optimization for rpc-enhancement module](https://github.com/Tencent/spring-cloud-tencent/pull/526)